requirements.txt
# Core HTTP and parsing
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.9.0  # for faster BS4 parsing

# Search indexing
whoosh>=2.7.4

# Browser automation (for FireCrawl)
selenium>=4.11.0
undetected-chromedriver>=3.5.0
webdriver-manager>=4.0.0

# HTML processing
html2text>=2020.1.16
trafilatura>=1.6.1

# Utility
python-dateutil>=2.8.2
tqdm>=4.65.0  # for progress bars
aiohttp>=3.8.5  # for async requests

# Additional requirements
python-dotenv>=1.0.0
warcio
dnspython
google-cloud-language
openai>=1.12.0
google-generativeai
azure-ai-textanalytics
PyPDF2>=3.0.0
anthropic>=0.18.1
ratelimit>=2.2.1
mechanize>=0.4.5 